% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRegress_Federation.R
\name{LRegress_Federation}
\alias{LRegress_Federation}
\title{Joins aggregates from 'N' nodes to compute a linear regression. We have to compute beta from the linear equation:
y = A*beta + error, but data y is distributed over local nodes.}
\usage{
LRegress_Federation(betas, Sigmas, Ndegree)
}
\arguments{
\item{betas}{regression coefficients computed in different 'N' nodes (list data type).}

\item{Sigmas}{Covariance matrices of the regression coefficients in different 'N' nodes (list data type).}

\item{Ndegree}{Number of degree of freedom.}
}
\value{
betaf  : Regression coefficient joining properly betas and Sigmas from 'N' nodes.
                 This will be the result the user will receive. The intermediate betas are not shown.
        Sigmaf : Covariance matrix of joining properly the partial covariance matrices coming from different nodes.
        t      : T-students
        pvals  : p-values, significance of the regression coefficients.
}
\description{
Joins aggregates from 'N' nodes to compute a linear regression. We have to compute beta from the linear equation:
y = A*beta + error, but data y is distributed over local nodes.
}
\keyword{regression}

