% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRegress_Node.R
\name{LRegress_Node}
\alias{LRegress_Node}
\title{Computes the linear regression.  We have to compute beta from the linear equation: y = A*beta + error
This is the function that will be running at local node level. The data (input parameters: y, A) are obtained from the
local databases using a specific queries. These queries will be the same for all nodes.}
\usage{
LRegress_Node(data, varname, covarnames, groups)
}
\arguments{
\item{data}{Data obtained from specific query to the local Database.}

\item{varname}{Name of the varname (column) that will dependent varname (left hand of the Linear model equation).}

\item{covarnames}{List of names of the co-varnames for the design or regression matrix that will be obtained  from specific query to the local Database.}

\item{groups}{List of names of the varnames to group, default to empty list}
}
\value{
betai  (rout[1]) : regression coefficient computed in nodes 'i'.
        Sigmai (rout[2]) : Covariance matrix of the regression coefficients betai.
                rout[3]  : Summary of the linear regression results.
}
\description{
Computes the linear regression.  We have to compute beta from the linear equation: y = A*beta + error
This is the function that will be running at local node level. The data (input parameters: y, A) are obtained from the
local databases using a specific queries. These queries will be the same for all nodes.
}
\keyword{regression}

